add support for MathJax v3
This bumps the plugin's version number to 2.0.

There's a "MathJax major version" setting, which is either 2 or 3.

The config and preamble are inserted appropriately according to the
version, and there are new defaults for version 3.

The default config for both versions now adds 'editor-rich-text' to the
list of ignored classes so MathJax doesn't run over the block editor.

I've rewritten the settings page to use add_settings_field, and store
all the options in one wp_options record, 'simple_mathjax_options'. I
think this is the preferred method, and it feels tidier.

fixes #22